requirements.txt
aiogram==2.19
Django==4.0.2
pyaml==21.10.1